Switch to using a single trendchart component across report and dashboard
