+++ app/js/components/trend-chart.vue
@@ -1,173 +1,177 @@
 <template lang="html">
-  <div v-show="metric.years.length > 1" class="qol-chart mdl-color--white mdl-shadow--2dp mdl-cell mdl-cell--4-col mdl-cell--12-col-tablet mdl-cell--12-col-desktop mdl-typography--text-center">
+  <div v-if="years.length > 1"
+       :class="framework === 'mdl' ? 'qol-chart mdl-color--white mdl-shadow--2dp mdl-cell mdl-cell--4-col mdl-cell--12-col-tablet mdl-cell--12-col-desktop mdl-typography--text-center' : 'qol-chart'">
     <div class="trendchart">
-      <h1 v-if="metric.config">{{ metric.config.title }}</h1>
+      <h1 v-if="metricConfig">{{ metricConfig.title }}</h1>
       <span class="legend"><svg class="icon legend-county"><use href="#icon-trending_up"/></svg> County</span>
-      <span v-show="selected.length > 0" class="legend"><svg class="icon legend-selected"><use href="#icon-trending_up"/></svg> Selected</span>
-      <div class="ct-trendchart"/>
+      <span class="legend"><svg class="icon legend-selected"><use href="#icon-trending_up"/></svg> Selected</span>
+      <div :id="'ct-trendchart-' + metricConfig.metric" class="ct-trendchart"/>
     </div>
   </div>
 </template>
 
 <script>
-import { mapState } from 'vuex';
 import Chartist from 'chartist';
-import isNumeric from '../modules/isnumeric';
-import { calcValue } from '../modules/metric_calculations';
-import { legendLabelNumber, prettyNumber } from '../modules/number_format';
+import { abbrNum, round, prettyNumber } from '../modules/number_format';
 
 require('../modules/chartist.axis.title.js');
 require('../modules/chartist.tooltip.js');
 
 export default {
-  name: 'Trendchart',
-  computed: mapState(['metric', 'selected']),
+  name: 'TrendChart',
+  props: {
+    years: {
+      type: Array,
+      required: true,
+    },
+    values: {
+      type: Object,
+      required: true,
+    },
+    countyValues: {
+      type: Object,
+      required: true,
+    },
+    metricConfig: {
+      type: Object,
+      required: true,
+    },
+    // Key to identify if this component needs to have MDL framework styling (used in the dashboard) or not.
+    framework: {
+      type: String,
+      required: false,
+      default: '',
+    },
+
+    // TODO: handle selected values.
+  },
+  computed: {
+    // The two "chart" computed variables filter the metric objects into arrays keyed by year, leaving null gaps where there's no data for a given year.
+    countyValuesChart() {
+      return this.years.map((year) => {
+        if (this.countyValues && this.countyValues.hasOwnProperty(year)) return this.countyValues[year];
+        return null;
+      });
+    },
+    valuesChart() {
+      return this.years.map((year) => {
+        if (this.values && this.values.hasOwnProperty(year)) return this.values[year];
+        return null;
+      });
+    },
+  },
   watch: {
-    'metric.data': 'renderChart',
-    'selected': 'renderChart',
+    countyValuesChart() { this.renderChart(); },
+  },
+  mounted() {
+    this.renderChart();
+  },
+  updated() {
+    this.renderChart();
   },
   methods: {
     renderChart() {
-      if (this.metric.years.length > 1) {
-        const _this = this;
-        const data = this.updateData();
-
-        // Set low and high values of axes based on overall data set.
-        const low = Object.values(this.metric.data.map).reduce((previous, val) => {
-          if (previous === null) {
-            return Math.min(...Object.values(val).filter(isNumeric));
-          }
-          return Math.min(previous,
-            ...Object.values(val).filter(isNumeric),
-          );
-        }, null);
-
-        const options = {
-          fullWidth: true,
-          height: '180px',
-          showArea: false,
-          low,
-          lineSmooth: Chartist.Interpolation.cardinal({
-            fillHoles: true,
-          }),
-          axisY: {
-            labelInterpolationFnc: (value, index) => (legendLabelNumber(value, _this.metric.config)),
+      // Render chart, but only if all the data is here (so render only once).
+      // Filter i=>i call returns true only if the array contains at least one non-null value.
+      if (this.years.length <= 1) return;
+      const metricConfig = this.metricConfig;
+      const len = this.years[this.years.length - 1] - this.years[0];
+      const options = {
+        fullWidth: true,
+        height: '180px',
+        showArea: false,
+        low: 0,
+        chartPadding: {
+          right: 40,
+        },
+        lineSmooth: Chartist.Interpolation.cardinal({
+          fillHoles: true,
+        }),
+        axisY: {
+          labelInterpolationFnc(value, index) {
+            return abbrNum(round(Number(value), 2), 2);
           },
-          plugins: [
-            Chartist.plugins.tooltip({
-              transformTooltipTextFnc (value) {
-                return prettyNumber(value, _this.metric.config.decimals, _this.metric.config.prefix, _this.metric.config.suffix, _this.metric.config.commas);
-              },
-            }),
-          ],
-        };
-        // axis labels
-        if (_this.metric.config.label) {
-          options.plugins.push(Chartist.plugins.ctAxisTitle({
-            axisX: {
-              axisTitle: '',
-              axisClass: 'ct-axis-title',
-              offset: {
-                x: 0,
-                y: 50,
-              },
-              textAnchor: 'middle',
+        },
+        axisX: {
+          labelInterpolationFnc(value, index) {
+            if (len > 6) {
+              return index % 2 === 0 ? value : null;
+            }
+            return value;
+          },
+        },
+        plugins: [
+          Chartist.plugins.tooltip({
+            transformTooltipTextFnc(value) {
+              return prettyNumber(value, metricConfig.decimals, metricConfig.prefix,
+                metricConfig.suffix);
             },
-            axisY: {
-              axisTitle: _this.metric.config.label,
-              axisClass: 'ct-axis-title',
-              offset: {
-                x: 0,
-                y: -1,
-              },
-              flipTitle: false,
-              textAnchor: 'middle',
-            },
-          }));
-        }
-
-        this.chart = new Chartist.Line('.ct-trendchart', data, options);
-        // animation
-        this.chart.on('draw', (data) => {
-          if (data.type === 'line') {
-            data.element.animate({
-              d: {
-                begin: 500 * data.index,
-                dur: 500,
-                from: data.path.clone().scale(1, 0).translate(0, data.chartRect.height()).stringify(),
-                to: data.path.clone().stringify(),
-                easing: Chartist.Svg.Easing.easeOutQuint,
-              },
-              opacity: {
-                begin: 500 * data.index,
-                dur: 500,
-                from: 0,
-                to: 1,
-              },
-            });
-          }
-        });
-      } else if (this.chart) {
-        this.chart.detach();
-        this.chart = null;
-      }
-    },
-    updateData() {
-      const metric = this.metric;
-      const chartData = {
-        labels: metric.years,
-        series: [],
+          }),
+        ],
       };
 
-      // county values
-      const keys = Object.keys(metric.data.map);
-      const areaArray = [];
+      // Axis labels
+      if (metricConfig.label) {
+        options.plugins.push(Chartist.plugins.ctAxisTitle({
+          axisX: {
+            axisTitle: '',
+            axisClass: 'ct-axis-title',
+            offset: {
+              x: 0,
+              y: 50,
+            },
+            textAnchor: 'middle',
+          },
+          axisY: {
+            axisTitle: metricConfig.label,
+            axisClass: 'ct-axis-title',
+            offset: {
+              x: 0,
+              y: -1,
+            },
+            flipTitle: false,
+            textAnchor: 'middle',
+          },
+        }));
+      }
 
-      // county value
-      for (let i = 0; i < chartData.labels.length; i++) {
-        let areaValue = null;
-        if (metric.years.indexOf(chartData.labels[i].toString()) !== -1) {
-          if (metric.config.world_val && metric.config.world_val[`y_${chartData.labels[i]}`]) {
-            areaValue = metric.config.world_val[`y_${chartData.labels[i]}`];
-          } else {
-            areaValue = calcValue(metric.data, metric.config.type, chartData.labels[i], keys);
-          }
+      const chart = new Chartist.Line(`#ct-trendchart-${metricConfig.metric}`, {
+        labels: this.years,
+        series: [this.valuesChart, this.countyValuesChart],
+      }, options);
+
+      // Animation.
+      chart.on('draw', (data) => {
+        if (data.type === 'line') {
+          data.element.animate({
+            d: {
+              begin: 500 * data.index,
+              dur: 500,
+              from: data.path.clone().scale(1, 0).translate(0, data.chartRect.height()).stringify(),
+              to: data.path.clone().stringify(),
+              easing: Chartist.Svg.Easing.easeOutQuint,
+            },
+            opacity: {
+              begin: 500 * data.index,
+              dur: 500,
+              from: 0,
+              to: 1,
+            },
+          });
         }
-        areaArray.push({
-          meta: 'County',
-          value: areaValue,
-        });
-      }
-      chartData.series.push(areaArray);
-      // selected values
-      if (this.selected.length > 0) {
-        const selectedArray = [];
-        for (let i = 0; i < chartData.labels.length; i++) {
-          if (metric.years.indexOf(chartData.labels[i].toString()) !== -1) {
-            const selectedValue = calcValue(metric.data, metric.config.type, chartData.labels[i], this.selected);
-            selectedArray.push({
-              meta: 'Selected',
-              value: selectedValue,
-            });
-          } else {
-            selectedArray.push(null);
-          }
-        }
-        chartData.series.push(selectedArray);
-      }
-      return chartData;
+      });
     },
   },
 };
 </script>
 
 <style lang="css">
-    .qol-chart .ct-series-b .ct-line,
-    .qol-chart .ct-series-b .ct-point {
-        stroke: #00688B;
-    }
     .qol-chart .ct-series-a .ct-line,
     .qol-chart .ct-series-a .ct-point {
+        stroke: #00688B;
+    }
+    .qol-chart .ct-series-b .ct-line,
+    .qol-chart .ct-series-b .ct-point {
         stroke: orange;
     }
     .ct-trendchart {
@@ -210,6 +214,10 @@ export default {
     .ct-area,
     .ct-line {
         pointer-events: none;
+    }
+    .ct-line {
+        fill: transparent;
+        stroke-width: 2px;
     }
 </style>
