+++ brunch-config.js
@@ -16,7 +16,7 @@ module.exports = {
       joinTo: {
         'vendor.js': /^(?!app)/,
         'app.js': ['app/js/main.js', 'app/js/Dashboard.vue', 'app/js/vuex-store/index.js', /^app\/js\/modules/, /^app\/js\/components\/(?!report)/],
-        'report.js': ['app/js/report.js', /^app\/js\/modules/, /^app\/js\/components\/report/],
+        'report.js': ['app/js/report.js', 'app/js/components/trend-chart.vue', /^app\/js\/modules/, /^app\/js\/components\/report/],
       },
     },
     stylesheets: {
